[ADD] SPLT draft.

+++ include/peafowl/peafowl.hpp
@@ -1218,7 +1218,7 @@ public:
    * @return The result of the dissection from L2 to L7.
    */
   DissectionInfo dissectFromL2(const std::string& pkt,
-                               uint32_t timestamp,
+                               double timestamp,
                                ProtocolL2 datalinkType);
 
   /**
@@ -1230,7 +1230,7 @@ public:
    * @return The result of the dissection from L3 to L7.
    */
   DissectionInfo dissectFromL3(const std::string& pkt,
-                               uint32_t timestamp);
+                               double timestamp);
 
   /**
    * Dissects the packet starting from the beginning of the L4 (UDP or TCP)
@@ -1243,7 +1243,7 @@ public:
    * @return The result of the dissection from L3 to L7.
    */
   DissectionInfo dissectFromL4(const std::string& pkt,
-                               uint32_t timestamp,
+                               double timestamp,
                                const DissectionInfo& info);
 
   /**
@@ -1266,7 +1266,7 @@ public:
    * @return The result of the L3 dissection.
    */
   DissectionInfo dissectL3(const std::string& pkt,
-                           uint32_t timestamp);
+                           double timestamp);
 
   /**
    * Extracts from the packet the L4 information.
@@ -1280,7 +1280,7 @@ public:
    * @return  The result of the L4 dissection.
    */
   DissectionInfo dissectL4(const std::string& pkt,
-                           uint32_t timestamp,
+                           double timestamp,
                            const DissectionInfo& info,
                            FlowInfoPrivate &flowInfoPrivate);
